Add a Kubernetes backend
This uses a Deployment Object[1] + a Service Object[2] to
run a webservice. The Deployment object will spawn a
ReplicaSet[3] object, which will spawn the actual pods running
a container image that is specific to the environment specified.
The container image will read code from NFS and listen always
on port 8000 (no clashes because containers!). The service
object will be used to route requests to these pods, and also
expose them as tools.wmflabs.org/toolname. The exposing is
achieved by the `kube2proxy` script, which watches for
services with the label 'tools.wmflabs.org/webservice' set to
'true'.

It also does a bunch of refactoring to take out state from
the webservice object that should really be in the backends,
and moves it appropriately.

We'll use the pykube api client for now, and bundle it in
with the submodule + symlink into the deb.

[1]: http://kubernetes.io/docs/user-guide/deployments/
[2]: http://kubernetes.io/docs/user-guide/services/
[3]: http://kubernetes.io/docs/user-guide/replicasets/

TODO:
  - Add more / better documentation
  - Fix up caching in pykube to make status checks faster

Change-Id: I2c578141836ca39e60ed1c69aed4beaadbc8c891