debian: Explictly manage python dependencies
We live in a wacky world where we are trying to support precise, trusty and
jessie hosts with a single package. We are also using build tooling that
doesn't exist in the precise universe. This combination leads to bad
automatically computed dependencies for generated deb packages due to minor
version differences between the trusty hosts that can build the package from
source and the precise hosts where we would like to install. I couldn't find
any method of hinting to the dh-python tooling that fills in the
${python:Depends} macro that we really are ok with any python 2.7.x revision
as a runtime dependency.

Bug: T145326
Change-Id: I87976fc0bf965cb98f1321c33bcd52b7e3b31e45