Route all logs to /dev/null
Let's do this rather than make it configurable, since doing config
via puppet in kubernetes land is kinda not the easiest. So our
workflow will be:

GridEngine:
 - Push new deb to aptly
 - Make sure new deb is in all webgrid hosts
 - Restart all webservices

Kubernetes
 - Build new kubernetes images
 - Pull on all worker nodes
 - Restart all webservices

To revert, we'll prepare another patch and another deb in advance,
and just do the same as above with the newer deb.

Bug: T149946
Change-Id: Ia2747b695e09efeba795ca4fcbf1482343487ed2