Wait for the Kubernetes pod to shut down after "stop"
Currently, with the Kubernetes backend "webservice stop" only waits
until the corresponding service and deployment have been deleted, but
not the pod itself.  If a webservice is started immediately
afterwards, for example by "webservice restart", this means that for a
short while two pods can exist, one for the terminating and one for
the starting webservice.  This causes the script to fail because it
cannot cope with two parallel webservice pods, and there are probably
other code paths that do not expect this.  In addition, this behaviour
can be surprising to users and administrators because the SGE backend
waits for the webservice job to be gone.

This change amends the logic of "webservice stop" for Kubernetes
backends so that it wait until the webservice pod has been deleted.

Bug: T156626
Change-Id: I51a38c5a6c096f269669e8fdfd3ac19af24ec20b