Mount & load project name dynamically from /etc/wmcs-project
This is for toolsbeta compatibility. Using the content of the file,
we determine the project, and therefore the correct prefix that the
tool account name uses. The file is also mounted to the pod so the
command stays usable inside the pod.

Also fixed two flake8 errors regarding E722 do not use bare except

Bug: T190893
Bug: T192244
Depends-On: Icb27ee91df965b580b7223436878514e7e8c417d
Change-Id: I2a5e66a9ca982538b2a2c7c7ca62e7c47f30db64