Make Kubernetes the default backend and warn when guessing
The most basic documentation on starting a webservice is just to run
`webservice start`. Switching the default backend to Kubernetes should
help us keep new users from accidentally starting on the grid and
instead nudge them into Kubernetes use.

The error message given when the selected backend does not match the
state of $HOME/service.manifest is also updated in this changeset. The
new error message attempts to be more descriptive of the problem and its
likely solutions.

Bug: T154504
Change-Id: I2eb6b3ebb6d4b11b605faf59f8988c271d611be1
Co-Authored-by: Bryan Davis <bd808@wikimedia.org>