Kubernetes: ignore terminating objects when searching
Kubernetes can be in the process of stopping one pod and starting
another with both pods matching the same set of label selector values.
This is most likely to happen when restarting a webservice that is
actively handling user requests or that has a slow shutdown process.
These Pods will eventually be removed either because the process
completes normally or because Kubernetes force kills the container. For
the purposes of our webservice wrapper we can ignore objects that
Kubernetes has told to shutdown even if they are not gone yet.

Bug: T156626
Change-Id: I6c1f19402d769b8972a3622fea8edf94d4e9eb72