Set custom mime-types
Lighttpd 1.4.40 made overriding keys in an existing array a hard error.
This in turn breaks the lame, but workable, workaround documented on
wikitech for correcting the charset of various static files.

We cannot easily change the behavior of lighttpd, but we can change the
default configuration that we generate for a webservice. The change made
here is to embed the mime-type mappings directly rather than generating
them at service startup. This allows us to modify the mappings to
include a `charset=utf-8` declaration for most `text/*` types and a few
other types. Doing this in our webservice code rather than via
a customized version version of create-mime.assign.pl should be easier
to maintain in the long term with multiple webservice runtime
environments (grid engine, kubernetes) and operating systems (Jessie,
Stretch, ...) to deal with.

These settings were generated with create-mime.conf.pl@9b7a32e from
https://github.com/lighttpd/lighttpd1.4. Additional charset=utf-8 were
manually added to "text/*" entries and mappings for
"application/vnd.openxmlformats-officedocument.*" were removed.

Bug: T178601
Bug: T215683
Change-Id: Ied455843fcbb50ca08adbcac7546658d30ce07ae