new k8s: Fix ingress object and enable toolsbeta ingress creation
The other thing here is that I realized that cascading deletes don't
work on the deprecated API versions of these objects, unfortunately. I am
fairly sure we are stuck with those until we leave the pykube framework
behind.

This should fix everything on this end to enable migration to the new
cluster.  We still need a newer kubectl package installed (puppet patch
coming) and testing around what this does for the "shell" command since
that uses an absolute path to kubectl (which we may want to change).

Bug: T236202
Bug: T214513
Bug: T215531
Change-Id: Ie6c45440fb9c9f67806b9d58edb5c9d1fc4ac839