k8s: Set default requests for the new cluster
For some reason, the requests for higher limits result in a
guaranteed qos class (limit and requests are the same) for pod resources.
This is not the intended behavior at this time.

Also cleaned up black formatting and adjusted the tox runner so that
you can run it locally without it trying to check everything in a venv.

Bug: T236202
Bug: T228499
Change-Id: Ia940d2d48f997a0c0b33410fb94c44e27e21aac9