kubernetes: resource requests should be proportional to limits
Since a request is ultimately for the scheduler, setting a very low
request on every container (even ones with large limits), oversubscribes
nodes severely.  Setting the request to something proportional to the limit
allows the scheduler to share resources better and also provides motivation
for users to "right size" their webservice instead of requesting the max since
the cluster will not schedule something that has too large a request that
cannot be placed in the current size cluster.

In local testing I found that decimal numbers easily replace the various
formats that Kubernetes considers acceptable. Therefore a simple expression
should be all that is needed to assign a proportional request.

Bug: T244289
Change-Id: Iad2844d03c8fcaea4c58f4a53dcb8205499089b1