Reuse toolforge.common.tool.PROJECT in KubernetesBackend
We ended up with two different places in the code doing the same action
of reading /etc/wmcs-project to determine which project the webservice
script is being called from. A bit of editing can solve that.

Also fix a bug waiting to happen where the 'type' argument to Backend
constructors was shadowing the built-in global function 'type'.

Change-Id: I69aab87539bf1683f64f09e9d69da281288e1852