Introduce command "template" feature
The $HOME/.webservicerc file introduced in 91670a1 was intended to help us
deal with the growing number of arguments to the webservice command by
making it easy to state that the Kubernetes backend should be used. In the
years since we have added even more arguments and that does not seem
likely to stop as we attempt to expose more control of Kubernetes
deployments.

During the 2020 Kubernetes cluster migration we had a temporary
`migrate` action which made it possible to reuse a tool's
service.manifest file across restarts. This was very handy, and it
inspired bd808 to think about a reasonable way to make this
functionality available for the 'start' action more generally.

This patch introduces the concept of a "template" file which can be used
to store arguments (and eventually other structured content) for
starting a webservice. The code will look for a "--template=..." cli
argument and also fallback to looking for a $HOME/service.template file.
The latter is what most tools will be expected to use, but we may find
interesting uses for multiple templates in a single tool as well.

A webservice template file is a YAML document. It can contain these
settings:
* backend: the backend to use (equivalent to --backend=...)
* cpu: the CPU reservation to ask for on Kubernetes (--cpu=...)
* mem: the memory reservation to ask for on Kubernetes (--mem=...)
* type: the type of webservice to start (TYPE)
* extra_args: extra arguments to pass to the backend

By saving desired startup state in a file, the user can use simple
`webservice stop; webservice start` commands again! This is already
a great help for tools using a TYPE which processes extra_args values.
It is expected to become more useful for other TYPEs as additional cli
arguments are exposed to control things like Kubenetes pod scaling and
ingress customizations.

Change-Id: Idc32b519be12be140e3c3a7e693c18004c3defa3