Add support for redirecting to toolforge.org
Introduce a "--canonical" cli argument which can be used to configure
a webservice to redirect traffic received at the legacy
https://tools.wmflabs.org/$TOOL/... URL to the matching canonical
https://$TOOL.toolforge.org URL.

For the Kubernetes backend, this is accomplished by configuring the
legacy Ingress with a "nginx.ingress.kubernetes.io/permanent-redirect"
annotation.

For the Grid Engine backend, configure webservice-runner to send the
desired canonical hostname to the proxy layer when registering the
webservice. Proxylistener and urlproxy will be updated to consume this
new information and handle requests appropriately.

Bug: T234617
Change-Id: Id71c385bd310d39498f04d87449de7abc6228b09
Depends-On: I5f06ff9fe682f782cbfd7f722fd44406a842e2bd