Replace pykube with a custom API client
The pykube library we have been using has been abandoned upstream
(<https://github.com/kelproject/pykube>). There is a forked version that
is continuing to add support for newer Kubernetes versions
(<https://pypi.org/project/pykube-ng/>), but it is not packaged for
Debian. There has also been a long standing task to get rid of pykube in
our codebase (T159892), so the pykube-ng route does not seem optimal.

After some discussion in T197930, we decided to at least temporarily try
rolling our own client code. This turned out to have a few tricky bits,
but generally was as easy as was hoped.

Bug: T197930
Change-Id: I41bb3e3ce8f3ecbb18cccfa738896ac03bec00d5