args: A few fixups
This changes the ConfigParser import to use the backported configparser,
which is installed on all bastions, from 3.2. It is an improvement over
the native py2 one and will fix the except clause which was broken
previously. The previous version assumed it might not be available.

More importantly, I've reverted the change to extra_args back to using
"*" as nargs because argparse.REMAINDER causes non-intuitive behaviors.

This also removes the unused import of subprocess.

Bug: T249390
Change-Id: Id99087327c896958a8d674ff76f387ddf7278534