Remove validation of Kubernetes self-signed API cert
In 9d93e9d we added a homegrown Kubernetes API client. One of the
tricker bits it handled was getting the requests library setup so that
it could verify the self-signed TLS certificate that the Kubernetes API
uses. This required creating a physical temporary file as requests
expects to be able to use os.path.exists() on the provided path. Under
python2.7 on the bastions we have seen that the __del__ implementation
to rm this file on destruction of the K8sClient object does not fire
reliably. This in turn led to a lot (like really a lot) of leaked files
in /tmp.

We considered shipping the signing cert bundle directly in the
application. This would however need to account at least for variations
between tools and toolsbeta, and it would be pretty much impossible to
deal with any local testing environment's needs.

We also considered provisioning the cert in some well known location via
Puppet, but that ends up also having issues for local testing and would
not be idea for in-container operations.

Ultimately we decided that the risk of man-in-the-middle attacks on the
TLS session between wherever webservice runs and the Kubernetes API
endpoint it is targeting is low and turning off requests' validation is
easy.

Bug: T253412
Change-Id: Ic10006edaecd364fbdf67694a1c1b28220cfa401