Disable tool name alias in lighttpd config with --canonical
When tools are using the legacy tools.wmflabs.org/$TOOL URLs, lighttpd
needs a `alias` directive to treat the tool's document root
($HOME/public_html) as the /$TOOL root path. When using the new
$TOOL.toolforge.org URLs, this alias is no longer needed. Leaving it in
place so that the lighttpd config was the same across both seemed fine,
but we have discovered that it causes problems serving static files
which begin with $TOOL.

This change makes all WebService sub-classes aware of the `--canonical`
flag. When using canonical mode, LighttpdPlainWebService and its sub
classes will comment out the `alias` directive in the generated lighttpd
config.

Bug: T254640
Change-Id: Ia5969d737d20baf6e9ec88af1c53b65a99e42d9f