Pass `--canonical` to webservice-runner inside k8s pod
Follow up to 17916c1 which actually passes the `--canonical` flag all
the way through to webservice-runner when using the Kubernetes backend.
This was not needed in prior usage of `--canonical` with Kubernetes as
the ingress routing there is managed using Ingress objects and not
runtime callbacks to dynamicproxy. With the new usage however it is
needed so that lighttpd based runtimes can change generated
configuration inside the container.

Bug: T254640
Change-Id: Ifb2c7750aa9621cd77ec968a13a79e44369c484f