Make `webservice shell` scriptable
Change the launch process for pods used by `webservice shell` so that
extra command line options can be passed through to the main container
in the pod as its pid 0 process if desired. Example: `webservice
python3.7 shell -- webservice-python-bootstrap --fresh` will start
a pod with `webservice-python-bootstrap --fresh` as pid 0 and return the
exit status of that process when it completes.

Our old solution was to start a pod using the Kubernetes API and then
attach to that running container after polling shows it running using
`kubectl attach`. The new system uses `kubectl run` instead which puts
kubectl in charge of the whole pod lifecycle. In initial testing this
also seems to be slightly faster to start the interactive shell, likely
because kubectl is doing fancier things under the hood to find out when
the pod is ready.

Another behavior change is that concurrent invocations of `webservice
shell` in the same tool/namespace will no longer result in multiple
consoles attached to the same running pod. This prior behavior was not
documented anywhere as far as I know, and also seemed to be surprising
to anyone who ran across it in the live system. The new behavior will be
to start a distinct pod for each invocation. This would have been a bit
scary on the old Kubernetes cluster, but now that we have per-namespace
quota restrictions the likelihood of run away pods consuming the entire
cluster seems low.

Bug: T169695
Change-Id: I8be6562acd674484704c162e7ef34f58781e2617