Look for service.template in various code directories
To encourage people to commit their service.template files to git (which
will be helpful for buildpacks), look for them in the various directories
that people are likely to directly clone their repository to, such as
~/www/python/src, ~/www/js, and ~/public_html.

Because it can now be in multiple places, error out if we find multiple
templates. For people who have already committed it to Git and symlink it
to ~/ (like me!), this will be a little bit of breakage and require cleanup,
but I think the advantage to everyone else should outweigh that.

Bug: T266692
Change-Id: I0ce894d1b8ac1f1bb8769918b62465abba822d63