gridengine: cleaner release default
Make the default release a bit cleaner and easier for maintenance. The parser
can handle the default and the available choices, no need to double check that
later. Improve the help message as well.

And while at it, pass the release argument to the gridengine backend, so what
the user specified can actually make it to qsub.

Bug: T278748
Change-Id: I217d231d415cd1bcf57e52abffc8d26d1461a544
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>