tools-webservice: add basic README file
People staring at tools-webservice for the first time may have a difficult
time understanding how it works, or how to setup a development environment,
or how are changes to it are meant to be done.

Introduce a basic README file as stop-gap, reflecting today's reality. We
can work later on making the development workflow for tools-webservice
easier.

Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: Ib8f1f237a9912629e4788dbc6d5a93a717a611cf