k8s: don't set command in the container if empty
When restarting, we check if the deployment has changed, and if so, we
recreate it as that will trigger the recreation of the pods too.

But when fetching the deployment definition from the api, k8s will
skip the command if it's empty, but we compare with the generated
spec, that has the command key set to empty list for buildservice
types, giving a false positive, and the deployment recreation just
does nothing.

This skips adding the command it it's empty, allowing us to detect if
the deployment has not changed, and removing the pods if it did not in
the case of a restart.

Bug: T335888
Signed-off-by: David Caro <dcaro@wikimedia.org>
Change-Id: I0830a83092796471d13a3e545bdf8376cd01b15d