Restart Kubernetes webservices more cleanly
Instead of deleting the running pod and then waiting for Kubernetes to
recreate it, let’s have Kubernetes create a new pod, wait for it to be
ready, switch traffic to it, and only then stop the old pod. To do this,
we add a timestamp to the metadata of the pod spec in the deployment,
and always update the deployment (not only when resources changed). When
the deployment is updated, Kubernetes will notice the existing pod no
longer matches the spec, and start the graceful restart process.
(`kubectl rollout restart deployment` uses the same trick, patching a
timestamp into the pod spec.)

This simplifies the restart itself (always update the deployment,
instead of sometimes updating it and sometimes deleting a pod inside it)
and removes a significant chunk of code to compare deployments, at the
cost of slightly complicating the check afterwards whether the restart
is finished (we now need to be able to distinguish between the old and
new pod, but the new startedAt annotation gives us exactly the
information we need for this). For most tools, the effect of this will
be to reduce, but not eliminate, the downtime during a restart.
However, tools that add startup, readiness or liveness probes – either
by manually patching their deployments or via a prospective future
webservice feature – can now even have restarts with no downtime:
Kubernetes will only remove the old pod once the new one is fully
started up and ready to accept traffic.

Bug: T337182
Change-Id: I2f3d2e7fa86f233d8db478531a0f8493d21cfbf4